move all Dynomite related files into the dynomite package
Some files had historical file locations that did not map well to DM
concepts. For example, sidecode/utils/DynomiteYamlTuneTask.java was not
a good location as this is not a generic utility, but instead is a
Dynomite specific task.
