+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/InjectedWebListener.java
@@ -31,7 +31,7 @@ import com.google.inject.name.Names;
 import com.google.inject.servlet.GuiceServletContextListener;
 import com.google.inject.servlet.ServletModule;
 import com.netflix.dynomitemanager.defaultimpl.DynomitemanagerConfiguration;
-import com.netflix.dynomitemanager.defaultimpl.DynomiteYamlTuner;
+import com.netflix.dynomitemanager.dynomite.DynomiteYamlTuner;
 import com.netflix.dynomitemanager.defaultimpl.IConfiguration;
 import com.netflix.dynomitemanager.dynomite.DynomiteProcessManager;
 import com.netflix.dynomitemanager.dynomite.IDynomiteProcess;
@@ -55,7 +55,7 @@ import com.netflix.dynomitemanager.sidecore.config.VpcInstanceDataRetriever;
 import com.netflix.dynomitemanager.sidecore.storage.IStorageProxy;
 import com.netflix.dynomitemanager.sidecore.storage.RedisStorageProxy;
 import com.netflix.dynomitemanager.sidecore.utils.FloridaHealthCheckHandler;
-import com.netflix.dynomitemanager.sidecore.utils.YamlTuner;
+import com.netflix.dynomitemanager.conf.YamlTuner;
 import com.netflix.dynomitemanager.supplier.HostSupplier;
 import com.netflix.dynomitemanager.supplier.CassandraHostsSupplier;
 import com.netflix.karyon.spi.HealthCheckHandler;
