+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/conf/package-info.java
@@ -1,6 +1,8 @@
 /**
  * Copyright 2016 Netflix, Inc.
  *
+ * Copyright 2016 DynomiteDB
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
  *
@@ -10,21 +12,15 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.dynomitemanager.sidecore.utils;
-
-import java.io.IOException;
 
 /**
- * Write a yaml file to disk.
+ * Dynomite Manager and module configuration. Configuration includes two distinct types of configuration:
+ *
+ * <ul>
+ * <li>Java properties: Properties are inbound configuration for Dynomite Manager, Dynomite, Redis, ARDB, etc. that are
+ * provided via Archaius.
+ * <li>Configuration files: Configuration files are outbound configuration used to configure processes under the control
+ * of Dynomite Manager, such as dynomite.yaml (Dynomite), redis.conf (Redis) and ardb.conf (ARD
+ * </ul>
  */
-public interface YamlTuner {
-
-    /**
-     * Generate and write a yaml file to disk.
-     *
-     * @param yamlLocation full path to a yaml file
-     * @throws IOException exception thrown if unable to write yaml file
-     */
-    void writeAllProperties(String yamlLocation) throws IOException;
-
-}
+package com.netflix.dynomitemanager.conf;
