+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/dynomite/DynomiteYamlTuneTask.java
@@ -10,17 +10,17 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.dynomitemanager.sidecore.utils;
+package com.netflix.dynomitemanager.dynomite;
 
 import java.io.IOException;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.netflix.dynomitemanager.defaultimpl.IConfiguration;
-import com.netflix.dynomitemanager.dynomite.DynomiteConfiguration;
 import com.netflix.dynomitemanager.sidecore.scheduler.SimpleTimer;
 import com.netflix.dynomitemanager.sidecore.scheduler.Task;
 import com.netflix.dynomitemanager.sidecore.scheduler.TaskTimer;
+import com.netflix.dynomitemanager.conf.YamlTuner;
 
 /**
  * Tune a process by writing a YAML configuration file.
