+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/dynomite/DynomiteYamlTuner.java
@@ -10,14 +10,14 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.dynomitemanager.defaultimpl;
+package com.netflix.dynomitemanager.dynomite;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import com.netflix.dynomitemanager.dynomite.DynomiteConfiguration;
+import com.netflix.dynomitemanager.defaultimpl.IConfiguration;
 import com.netflix.dynomitemanager.identity.InstanceIdentity;
 import com.netflix.dynomitemanager.sidecore.storage.IStorageProxy;
-import com.netflix.dynomitemanager.sidecore.utils.YamlTuner;
+import com.netflix.dynomitemanager.conf.YamlTuner;
 import com.netflix.dynomitemanager.IInstanceState;
 
 import org.slf4j.Logger;
