+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/dynomite/package-info.java
@@ -1,0 +1,20 @@
+/**
+ * Copyright 2016 Netflix, Inc.
+ *
+ * Copyright 2016 DynomiteDB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+/**
+ * Manage Dynomite including inbound configuration via Archaius, process management (start/stop), health checks, access
+ * to the Dynomite admin REST API, and outbound configuration via dynomite.yaml.
+ */
+package com.netflix.dynomitemanager.dynomite;
